requirements.txt
cycler==0.10.0
decorator==4.4.0
imageio==2.5.0
kiwisolver==1.1.0
matplotlib==3.1.0
networkx==2.3
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
pkg-resources==0.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
pygraphviz==1.3
